pyproject.toml
[tool.poetry]
name = "testlio"
version = "0.1.0"
description = ""
authors = ["DT"]

[tool.poetry.dependencies]
python = "3.8.12"
dvc = "^2.10.0"
hydra-core = "1.1.1"
pdoc3 = "^0.10.0"
notebook = "^6.4.10"
pandas = "^1.4.2"
scipy = "1.8.0"
scikit-learn = "1.0.2"
prefect = "1.2.0"
dirty-cat = "^0.2.0"
fuzzywuzzy = "^0.18.0"
python-Levenshtein = "^0.12.2"
pandas-profiling = "^3.2.0"
ipywidgets = "^7.7.0"
imbalanced-learn = "^0.9.0"
mlflow = "^1.25.1"
dagshub = "^0.1.8"
pre-commit = "^2.18.1"
xgboost = "^1.6.0"
hyperopt = "^0.2.7"
fastapi = "^0.76.0"
uvicorn = "^0.17.6"
pydantic = "^1.9.0"


[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^2.17.0"
plotly = "^5.7.0"
matplotlib = "^3.5.2"
pytest-steps = "^1.8.0"
pandera = "^0.11.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build   
)/ 
'''
